✍️  Write "Starting Over" post
I _know_ this will need to get moved into some directory, with a name
that is not "index"  as soon as it finds a companion, and they need to
be organized in some way.

But I'm trying to exercise the muscles of building something from
scratch. Not trying to predict the future of where the project will go,
but adapting to the needs as they actually arise.

Maybe this pragmatism will bite me. Time will tell!