🚧 Move "Starting Over" off of the root index
Well that didn't take long. What lived at `index.md` is really it's own
piece of writing, and probably should not occupy the root page. So I'm
moving it to it's own file.

Doing this makes for three items in the `collections.all` collection.
But the root index is not really _part_ of a collection, so it feels
like Eleventy's `eleventyExcludeFromCollections` fits the bill for this
page.