🗺 Render the list only on the index
I've noticed in my writing about my writing, that it's really hard to
avoid loaded language. "Posts" and "entries" keep being typed, and
deleted. "Homepage" vs "index". These may be mundane decisions, with few
implications, but I'm trying to defer them until I "have more
information". It's a bit of a pedantic practice... but that pedantry is
what I'm trying to explore in building from scratch.

The navigation is replaced with a "Back to the index" link. I could just
call this "homepage" and be done with it... but again the deferring of
naming things until I know what they are seems good, in theory.