🍣 Add a "blogroll" page
The great revelation of early HTML development was the lowly anchor tag.
You could link to anything. Personal websites linking to personal
websites was the best use of this tag.

There's so much power in this tag. It's the reason Instagram only gives
you one of them in your profile, and doesn't allow them in posts. It's
the reason the walled gardens have incentivized expanding links with
rich previews. The link is the building block of the Internet.

This is a brain dump of the personally published websites I draw
inspiration from. It will grow over time.