🐙 Generate a JSON structure from git commits
I have this idea that I'd like to surface the commits on the website in
some way or another. It seemed like a fun exercise to try it out.

The prior art found in GitHub gists got me headed in the right
direction. While the bash-only solutions were workable, I prefer to have
something I understand, and can debug should the time ever come.

Deciding a to use a higher level language brought me to another
decision. Double down on JavaScript in this 11ty site, or reach for the
expressiveness of Ruby? I went with Ruby. The overhead is that I'll have
to subshell from JavaScript to Ruby, and subshell from Ruby to the git
command. But what's one more, really?