⚙️  Make the git commits available as a custom collection
There were two approaches I considered for making the git commits
available within templates.

1. Loading it in a [data file][]
2. Defining a [custom collection][]

The biggest deciding factor was Markdown. Commits are authored with
Markdown conventions, and I'd like those to be translated into HTML on
the website. Configuring another Markdown-It parser in a data file
is definitely possible... but I've already got one defined right here in
the Eleventy config, so why not reuse it.

I could see going back to loading via data files if this starts to feel
goofy. The impact in templates will be minimal.

[data file]: https://www.11ty.dev/docs/data-global/
[custom collection]: https://www.11ty.dev/docs/collections/#advanced-custom-filtering-and-sorting