🚧 Add an "under construction" banner
This is an idea borrowed from [Eric Bailey][]'s personal website. It
establishes a helpful context when browsing the site, and I'd like the
same for mine.

It calls to mind the "under construction" websites of old. They wouldn't
have anything to show for their construction, but a placeholder was
launched for...who knows why. Fear of missing out on a lead, if the page
was completely blank, and banking on their return to perpetually check
if the construction was complete?

To stretch the analogy, this approach feels less like a fence keeping
visitors out of a construction zone, and more like a hard hat to wear
upon entry.

[eric bailey]: https://ericwbailey.design/