🗓 Give the commits page a date
It kept appearing at the top of the list, because the file creation time
is always the time of deploy. This is [a common pitfall][] in the
Eleventy docs.

Eventually, this page may not even appear in an index, and have a date
associated with it. For now, give everything a date.

[a common pitfall]: https://www.11ty.dev/docs/dates/#collections-out-of-order-when-you-run-eleventy-on-your-server