📝 Make "Today I Learned" content available on this site
In the few days I've spent iterating on this project, it's become clear
to me that I want this site to completely replace my existing website,
soon. I like the idea of building in the open, and the sooner I can
close things out on the old implementation, the better.

Bringing over the content so it's available by direct link only fits the
bill for now. `eleventyExcludeFromCollections: true` makes that super
simple to accomplish.