📈 Build that writing habit
Some type-as-you-think poetry emerged today. Most of it is garbage. "The
times they are a changelog" feels like a good note to end on. That
intersection of colloquialisms and developer nomenclature always makes
me happy. I know it's a narrow audience who will get it. The specific of
that mystery invite some in further, and the nonsense is lost on others.

Writing daily in git is interesting. You have first, the writing. Then a
commit message is the reflection on the writing. Wild.